Revert test feature to be more inline with rust testing standards
