+++ .vscode/settings.json
@@ -11,10 +11,6 @@
     "files.insertFinalNewline": true,
     "gitlens.showWhatsNewAfterUpgrades": false,
     "lldb.consoleMode": "evaluate",
-    "rust-analyzer.cargo.features": [
-        "test"
-    ],
-    "rust-analyzer.cargo.noDefaultFeatures": true,
     "rust-analyzer.check.command": "clippy",
     "rust-analyzer.runnables.extraTestBinaryArgs": [
         "--nocapture"
