+++ src/uniffi/mod.rs
@@ -2,6 +2,8 @@
 pub mod error;
 /// Components of the data model.
 pub mod model;
+/// Operators for pipeline
+pub mod operator;
 /// Interface into container orchestration engine.
 pub mod orchestrator;
 /// Data persistence provided by a store backend.
